major cleanup

+++ cst/egmm.py
@@ -4,7 +4,7 @@ EGMM - Empirical Ground Motion Model
 """
 import numpy as np
 
-def cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb ):
+def cbnga(T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb):
     """
     2008 Campbell-Bozorgnia NGA ground motion relation
 
@@ -33,14 +33,14 @@ def cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb ):
          ground motion parameters, Tech. Rep. PEER 2007/02, Pacific Earthquake
          Engineering Research Center.
     """
-    M     = np.array( M )
-    R_RUP = np.array( R_RUP )
-    R_JB  = np.array( R_JB )
-    Z_TOR = np.array( Z_TOR )
-    Z_25  = np.array( Z_25 )
-    V_S30 = np.array( V_S30 )
-    delta = np.array( delta )
-    lamb  = np.array( lamb )
+    M     = np.array(M)
+    R_RUP = np.array(R_RUP)
+    R_JB  = np.array(R_JB)
+    Z_TOR = np.array(Z_TOR)
+    Z_25  = np.array(Z_25)
+    V_S30 = np.array(V_S30)
+    delta = np.array(delta)
+    lamb  = np.array(lamb)
     params = {
     'T':   (  'c0', 'c1',  'c2', 'c3',  'c4','c5', 'c6','c7', 'c8','c9','c10','c11','c12',   'k1',  'k2','k3','slY','tlY','sT','rho'),
     0.010: ( -1715,  500,  -530, -262, -2118, 170, 5600, 280, -120, 490, 1058,  40,  610,  865000, -1186, 1839, 478, 219, 526, 1000),
@@ -69,7 +69,7 @@ def cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb ):
     'PGD': ( -5270, 1600,   -70,    0, -2000, 170, 4000,   0,    0,   0, -820, 300, 1000,  400000,     0, 2744, 667, 485, 825,  174),
     }
 
-    params = 0.001 * np.array( params[T] )
+    params = 0.001 * np.array(params[T])
     n  = 1.18
     cc = 1.88
     c  = params[:13]
@@ -77,45 +77,45 @@ def cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb ):
     sigma_lnY, tau_lnY, sigmaT, rho = params[16:]
     sigma_lnAF  = 0.3
     sigma_lnGPA = 0.478 #FIXME
-    sigma_lnY_B = np.sqrt( sigma_lnY ** 2 - sigma_lnAF ** 2 )
-    sigma_lnA_B = np.sqrt( sigma_lnGPA ** 2 - sigma_lnAF ** 2 )
+    sigma_lnY_B = np.sqrt(sigma_lnY ** 2 - sigma_lnAF ** 2)
+    sigma_lnA_B = np.sqrt(sigma_lnGPA ** 2 - sigma_lnAF ** 2)
 
     f_mag = (
         c[0] +
         c[1] * M +
-        c[2] * np.maximum( 0.0, M - 5.5 ) +
-        c[3] * np.maximum( 0.0, M - 6.5 )
+        c[2] * np.maximum(0.0, M - 5.5) +
+        c[3] * np.maximum(0.0, M - 6.5)
     )
-    f_dis = (c[4] + c[5] * M) * np.log( np.sqrt( R_RUP * R_RUP + c[6] * c[6] ) )
-    F_RV = np.zeros_like( lamb )
-    F_NM = np.zeros_like( lamb )
-    F_RV[ (  30 < lamb) & (lamb < 150) ] = 1.0
-    F_NM[ (-150 < lamb) & (lamb < -30) ] = 1.0
-    f_flt = c[7] * F_RV * min( 1.0, Z_TOR ) + c[8] * F_NM
+    f_dis = (c[4] + c[5] * M) * np.log(np.sqrt(R_RUP * R_RUP + c[6] * c[6]))
+    F_RV = np.zeros_like(lamb)
+    F_NM = np.zeros_like(lamb)
+    F_RV[(30 < lamb) & (lamb < 150)] = 1.0
+    F_NM[(-150 < lamb) & (lamb < -30)] = 1.0
+    f_flt = c[7] * F_RV * min(1.0, Z_TOR) + c[8] * F_NM
     i = (R_JB > 0.0) & (Z_TOR >= 1.0)
-    f_hng = np.maximum( R_RUP, np.sqrt( R_JB * R_JB + 1.0 ) )
+    f_hng = np.maximum(R_RUP, np.sqrt(R_JB * R_JB + 1.0))
     f_hng = (f_hng - R_JB) / f_hng
     f_hng[i] = (R_RUP[i] - R_JB[i]) / R_RUP[i]
-    f_hng = ( c[9] * f_hng *
-        np.minimum( 1.0, np.maximum( 0.0, 2.0 * M - 12.0 ) ) *
-        np.maximum( 0.0, 1.0 - 0.05 * Z_TOR ) *
-        np.minimum( 1.0, 4.5 - 0.05 * delta )
+    f_hng = (c[9] * f_hng *
+        np.minimum(1.0, np.maximum(0.0, 2.0 * M - 12.0)) *
+        np.maximum(0.0, 1.0 - 0.05 * Z_TOR) *
+        np.minimum(1.0, 4.5 - 0.05 * delta)
     )
-    f_site = (c[10] + k[1] * n) * np.log( np.minimum( 1100.0, V_S30 ) / k[0] )
+    f_site = (c[10] + k[1] * n) * np.log(np.minimum(1100.0, V_S30) / k[0])
     i = V_S30 < k[0]
-    lowvel = np.any( i )
+    lowvel = np.any(i)
 
     if lowvel:
-        sigmaT = sigmaT * np.ones_like( V_S30 )
-        V_1100 = 1100.0 * np.ones_like( V_S30 )
-        A_1100 = cbnga( 'PGA', M, R_RUP, R_JB, Z_TOR, Z_25, V_1100, delta, lamb )[0]
+        sigmaT = sigmaT * np.ones_like(V_S30)
+        V_1100 = 1100.0 * np.ones_like(V_S30)
+        A_1100 = cbnga('PGA', M, R_RUP, R_JB, Z_TOR, Z_25, V_1100, delta, lamb)[0]
         f_site[i] = (
-            c[10] * np.log( V_S30[i] / k[0] ) +
+            c[10] * np.log(V_S30[i] / k[0]) +
             k[1] * (
-                np.log( A_1100[i] + cc * (V_S30[i] / k[0]) ** n ) -
-                np.log( A_1100[i] + cc )
+                np.log(A_1100[i] + cc * (V_S30[i] / k[0]) ** n) -
+                np.log(A_1100[i] + cc)
             )
-        ).astype( f_site.dtype )
+        ).astype(f_site.dtype)
         alpha = k[1] * A_1100 * (
             1.0 / (A_1100 + cc * (V_S30 / k[0]) ** n) -
             1.0 / (A_1100 + cc)
@@ -124,21 +124,21 @@ def cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb ):
             tau_lnY ** 2 +
             sigma_lnY_B ** 2 +
             sigma_lnAF ** 2 +
-            ( alpha * sigma_lnA_B ) ** 2 +
+            (alpha * sigma_lnA_B) ** 2 +
             2.0 * alpha * rho * sigma_lnY_B * sigma_lnA_B
         )
-        sigmaT[i] = np.sqrt( sigmaT2[i] ).astype( sigmaT.dtype )
+        sigmaT[i] = np.sqrt(sigmaT2[i]).astype(sigmaT.dtype)
 
-    f_sed = np.zeros_like( Z_25 )
+    f_sed = np.zeros_like(Z_25)
     i = Z_25 < 1
     f_sed[i] = c[11] * (Z_25[i] - 1.0)
     i = Z_25 > 3
-    f_sed[i] = c[12] * k[2] * np.exp( -0.75 ) * (
-        1 - np.exp( -0.25 * (Z_25[i] - 3.0) ) )
-    Y = np.exp( f_mag + f_dis + f_flt + f_hng + f_site + f_sed )
+    f_sed[i] = c[12] * k[2] * np.exp(-0.75) * (
+        1 - np.exp(-0.25 * (Z_25[i] - 3.0)))
+    Y = np.exp(f_mag + f_dis + f_flt + f_hng + f_site + f_sed)
 
     if not lowvel:
-        sigmaT = sigmaT * np.ones_like( Y )
+        sigmaT = sigmaT * np.ones_like(Y)
 
     return Y, sigmaT
 
@@ -165,53 +165,53 @@ def test():
     delta = 90.0,
     lamb = 0.0,
 
-    M = np.arange( 4.0, 8.501, 0.1 )
-    Y, sigma = cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb )
-    plt.figure( 1 )
+    M = np.arange(4.0, 8.501, 0.1)
+    Y, sigma = cbnga(T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb)
+    plt.figure(1)
     plt.clf()
-    plt.plot( M, Y )
-    plt.xlabel( 'M' )
-    plt.ylabel( T )
+    plt.plot(M, Y)
+    plt.xlabel('M')
+    plt.ylabel(T)
     M = 5.5,
 
-    V_S30 = np.arange( 180.0, 1500.1, 10.0 )
-    Y, sigma = cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb )
-    plt.figure( 2 )
+    V_S30 = np.arange(180.0, 1500.1, 10.0)
+    Y, sigma = cbnga(T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb)
+    plt.figure(2)
     plt.clf()
-    plt.plot( V_S30, sigma )
-    plt.xlabel( '$V_{S30}$' )
-    plt.ylabel( T )
+    plt.plot(V_S30, sigma)
+    plt.xlabel('$V_{S30}$')
+    plt.ylabel(T)
     V_S30 = 760.0,
 
-    Z_25 = np.arange( 0.0, 6.01, 0.1 )
-    Y, sigma = cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb )
-    plt.figure( 3 )
+    Z_25 = np.arange(0.0, 6.01, 0.1)
+    Y, sigma = cbnga(T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb)
+    plt.figure(3)
     plt.clf()
-    plt.plot( Z_25, Y )
-    plt.xlabel( '$Z_{2.5}$' )
-    plt.ylabel( T )
+    plt.plot(Z_25, Y)
+    plt.xlabel('$Z_{2.5}$')
+    plt.ylabel(T)
     Z_25 = 1.0,
 
-    TT = ( 0.01, 0.02, 0.03, 0.05, 0.075, 0.1, 0.15, 0.2, 0.25, 0.3,
-           0.4, 0.5, 0.75, 1.0, 1.5, 2.0, 3.0, 4.0, 5.0, 7.5, 10.0 )
+    TT = (0.01, 0.02, 0.03, 0.05, 0.075, 0.1, 0.15, 0.2, 0.25, 0.3,
+           0.4, 0.5, 0.75, 1.0, 1.5, 2.0, 3.0, 4.0, 5.0, 7.5, 10.0)
     Y = []
     sigma = []
     for T in TT:
-        tmp = cbnga( T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb )
-        Y += [ tmp[0][0] ]
-        sigma += [ tmp[1][0] ]
+        tmp = cbnga(T, M, R_RUP, R_JB, Z_TOR, Z_25, V_S30, delta, lamb)
+        Y += [tmp[0][0]]
+        sigma += [tmp[1][0]]
 
-    plt.figure( 4 )
+    plt.figure(4)
     plt.clf()
-    plt.loglog( TT, Y )
-    plt.xlabel( 'T' )
-    plt.ylabel( 'SA' )
+    plt.loglog(TT, Y)
+    plt.xlabel('T')
+    plt.ylabel('SA')
 
-    plt.figure( 5 )
+    plt.figure(5)
     plt.clf()
-    plt.semilogx( TT, sigma )
-    plt.xlabel( 'T' )
-    plt.ylabel( '$\sigma$' )
+    plt.semilogx(TT, sigma)
+    plt.xlabel('T')
+    plt.ylabel('$\sigma$')
 
     plt.draw()
     plt.show()
